backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.2
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.128.1
greenlet==3.3.1
h11==0.16.0
idna==3.11
passlib==1.7.4
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pypdf==6.6.2
python-dotenv==1.2.1
python-multipart==0.0.22
qrcode==8.2
reportlab==4.4.9
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
